// server/src/generated/prisma/enums.ts
export const CandidateStatus = {
  active: 'active',
  inactive: 'inactive',
  deployed: 'deployed',
  blacklisted: 'blacklisted'
} as const

export type CandidateStatus = (typeof CandidateStatus)[keyof typeof CandidateStatus]
// ...
export const DeploymentStatus = {
  active: 'active',
  completed: 'completed',
  terminated: 'terminated',
  extended: 'extended'
} as const

export type DeploymentStatus = (typeof DeploymentStatus)[keyof typeof DeploymentStatus]